// src/codegen/grpc-gateway.ts
export interface InitReq extends RequestInit {
  pathPrefix?: string
}